* move dependency nodes to separate class + Compact after remove elements from THashSet
